feat: improve formatting of relative times (#62)
* Improve formatting of relative times

This improves the display of relative times.  With the current code, a operation that starts in 4 minutes 59 seconds from now is reported to start in "4 minutes". An operation starting in 1 day 23 hours is reported to start in "1 day". This leads to confusion about when the event actually starts.  Using Carbon's time rounding, we now display more intuitive relative distances.

Closes #49

* Hide end time when not set

This prevents listing ongoing events as "ended 1 second ago" when they are actually still active